refactor: extract mcp server tools
